[MODIFY] Name 앞 prefix 제거
